// src/app/interfaces/registerPay.ts
export interface RegisterPay {
    name: string;
    dni: string;
    email: string;
    nroContrato: string;
    img: string;
    bank: string;
    id_Cuba:string;
    amount: string;
    voucher: string;
    date: Date;
    nameTitular: string;
    dniTitular: string;
    emailTitular: string;
    note: string;
    serial: string;
}